catalog mixin
