+++ aqua/core/backend/__init__.py
@@ -1,9 +1,15 @@
+from .backend import Backend
+from .backend_intake_fdb import BackendIntakeFDB
+from .backend_intake_xarray import BackendIntakeXarray
 from .backend_factory import BackendFactory
 from .backend_xarray import BackendXarray
-from .backend_intake_fdb import BackendIntakeFDB
+from .catalog_mixin import CatalogMixin
 
 __all__ = [
+    "Backend",
     "BackendFactory",
-    "BackendXarray",
     "BackendIntakeFDB",
+    "BackendIntakeXarray",
+    "BackendXarray",
+    "CatalogMixin",
 ]
